requirements.txt
Django==1.9.5
Fabric==1.10.2
Jinja2==2.8
MarkupSafe==0.23
Pillow==2.8.2
PyYAML==3.11
SQLAlchemy==1.0.8
Scrapy==1.0.5
Twisted==15.5.0
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
autopep8==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.22
celery==3.1.20
certifi==2016.2.28
cffi==1.5.2
characteristic==14.3.0
cryptography==1.2.2
cssselect==0.9.1
decorator==4.0.9
django-appconf==1.0.2
django-celery==3.1.17
django-cors-headers==1.1.0
django-dbbackup==2.5.0
django-encrypted-fields==1.1.2
django-extensions==1.6.6
django-imagekit==3.2.6
django-push-notifications==1.4.1
django-rest-framework-mongoengine==3.3.0
django-storages==1.5.0
django-timedeltafield==0.7.10
django-user-accounts==1.3.1
djangocms-admin-style==1.1.0
djangorestframework==3.3.2
djorm-ext-pgfulltext==0.10
djrill==1.4.0
docutils==0.12
enum34==1.1.6
et-xmlfile==1.0.1
greenlet==0.4.7
gunicorn==19.3.0
httplib2==0.9.2
idna==2.1
inflect==0.2.5
ipaddress==1.0.16
ipython==4.0.1
ipython-genutils==0.1.0
jdcal==1.2
jsmin==2.2.0
juspay==1.1.1
kombu==3.0.35
lxml==3.6.0
meld3==1.0.2
mixpanel==4.3.0
mongoengine==0.10.6
openpyxl==2.3.3
paramiko==2.0.0
path.py==8.1.2
pdfkit==0.5.0
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.5
pika==0.10.0
pilkit==1.1.12
psycopg2==2.6.1
ptyprocess==0.5
pushbots==0.0.1
pyOpenSSL==0.15.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
pymongo==3.2.2
pyotp==2.0.1
python-dateutil==2.5.2
python-gcm==0.2
python-keyczar==0.715
python-memcached==1.57
pytz==2016.3
queuelib==1.4.2
redis==2.10.5
reportlab==3.2.0
requests==1.2.2
service-identity==14.0.0
simple-crypt==4.1.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.9.0
supervisor==3.3.0
tornado==4.3
tornado-redis==2.4.18
traitlets==4.0.0
w3lib==1.13.0
wsgiref==0.1.2
zope.interface==4.1.3
django-ses
seacucumber
google-api-python-client
elasticsearch
django-htmlmin
